完善日志和工具类
