+++ shadowsocks-ui/pom.xml
@@ -22,5 +22,10 @@
             <artifactId>json</artifactId>
             <version>LATEST</version>
         </dependency>
+        <dependency>
+            <groupId>org.slf4j</groupId>
+            <artifactId>slf4j-jdk14</artifactId>
+            <version>1.7.25</version>
+        </dependency>
     </dependencies>
 </project>
